+++ packages/ice-scripts/test/jest/alias/sum.spec.js
@@ -1,0 +1,5 @@
+import sum from '@/utils';
+
+test('alias', () => {
+  expect(sum(1, 2)).toBe(3);
+});