+++ packages/ice-scripts/test/jest/customConfig/sum.e2e.js
@@ -1,0 +1,5 @@
+import sum from './utils';
+
+test('jest config', () => {
+  expect(sum(1, 2)).toBe(3);
+});